# docs/swagger.yaml
basePath: /
definitions:
  dto.AuthResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        description: access-token lifetime in seconds
        type: integer
      refreshToken:
        type: string
      tokenType:
        description: always "Bearer"
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.UpdateRoleRequest:
    properties:
      role:
        enum:
        - admin
        - user
        type: string
    required:
    - role
    type: object
  dto.UpdateUserRequest:
    properties:
      birthdate:
        type: string
      firstName:
        maxLength: 100
        type: string
      gender:
        enum:
        - male
        - female
        type: string
      isActive:
        type: boolean
      lastName:
        maxLength: 100
        type: string
      middleName:
        maxLength: 100
        type: string
    type: object
  dto.UserResponse:
    properties:
      birthdate:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      lastName:
        type: string
      middleName:
        type: string
      role:
        type: string
    type: object
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
  response.Pagination:
    properties:
      hasNext:
        type: boolean
      hasPrevious:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
info:
  contact: {}
  description: REST API for the URL shortener service.
  title: URL Shortener API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies credentials and returns a JWT access/refresh token pair.
      parameters:
      - description: Login data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Log in
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access/refresh pair.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates an account and returns a JWT access/refresh token pair.
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Register a new user
      tags:
      - auth
  /users:
    get:
      description: Returns a paginated list of users.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft-deletes a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      description: Returns a single user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates a user's profile fields.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /users/{id}/role:
    put:
      consumes:
      - application/json
      description: Assigns a role (admin or user) to a user. Admin only.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Set a user's role
      tags:
      - users
  /users/me:
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's own profile.
      parameters:
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update own profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"